+++ s2gos-common/src/tests/process/cli/test_request.py
@@ -26,6 +26,41 @@ class RequestTest(unittest.TestCase):
         process_request = processing_request.as_process_request()
         self.assertIsInstance(process_request, ProcessRequest)
 
+    def test_as_process_request_dotpath(self):
+        processing_request = ProcessingRequest(
+            process_id="P16",
+            dotpath=True,
+            inputs={
+                "scene.colors.bg": "blue",
+                "scene.colors.opacity": 0.7,
+                "scene.colors.fg": "white",
+                "scene.models.path": "models/*",
+                "log_file": "logs/*",
+            },
+        )
+        process_request = processing_request.as_process_request()
+        self.assertIsInstance(process_request, ProcessRequest)
+        self.assertEqual(
+            {
+                "scene": {
+                    "colors": {
+                        "bg": "blue",
+                        "opacity": 0.7,
+                        "fg": "white",
+                    },
+                    "models": {
+                        "path": "models/*",
+                    },
+                },
+                "log_file": "logs/*",
+            },
+            process_request.inputs,
+        )
+
+    def test_create_processing_request_with_dotpath(self):
+        request = ProcessingRequest.create(process_id="x15", dotpath=True)
+        self.assertEqual(ProcessingRequest(process_id="x15", dotpath=True), request)
+
     def test_read_processing_request_from_yaml_stdin(self):
         stream = StringIO("process_id: test_func\ninputs:\n  x: 7\n  y: 9")
         with patch("sys.stdin", new=stream):
@@ -127,6 +162,7 @@ class RequestTest(unittest.TestCase):
         with pytest.raises(click.ClickException, match="Invalid request NAME: '2x'"):
             ProcessingRequest.create(process_id="my_func", inputs=["2x=20'"])
 
+    # noinspection PyMethodMayBeStatic
     def test_read_processing_request_from_invalid_subscription(self):
         with pytest.raises(
             click.ClickException,
@@ -157,3 +193,14 @@ class RequestTest(unittest.TestCase):
             ProcessingRequest.create(
                 process_id="my_func", subscribers=["failed=localhorst"]
             )
+
+
+class ProcessingRequestHelpersTest(unittest.TestCase):
+    def test_nest_dict(self):
+        self.assertEqual(
+            {"a": 1, "b": True}, ProcessingRequest._nest_dict({"a": 1, "b": True})
+        )
+        self.assertEqual(
+            {"a": 1, "b": {"x": 0.3, "y": -0.1}},
+            ProcessingRequest._nest_dict({"a": 1, "b.x": 0.3, "b.y": -0.1}),
+        )